refactor: hide provisional Rust API surface
