+++ Cargo.toml
@@ -28,6 +28,10 @@ exclude = [
 ]
 
 
+[features]
+fuzzing = []
+internal = []
+
 [package.metadata.binstall]
 pkg-url = "{ repo }/releases/download/v{ version }/{ name }-{ target }-v{ version }.tar.gz"
 bin-dir = "{ name }-{ target }-v{ version }/{ bin }{ binary-ext }"
@@ -44,10 +48,12 @@ path = "src/main.rs"
 [[bench]]
 name = "tachyonfx"
 harness = false
+required-features = ["internal"]
 
 [[bench]]
 name = "core"
 harness = false
+required-features = ["internal"]
 [dependencies]
 
 ratatui = { version = "0.30.2", default-features = false, features = ["crossterm", "layout-cache", "std"] }
