+++ src/lib.rs
@@ -1,3 +1,8 @@
+//! Excise's supported product is the `excise` command-line program.
+//!
+//! The library target shares private implementation modules with the binary. It is not a
+//! supported Rust API; use the CLI and its documented configuration and report contracts.
+
 #![allow(
     clippy::unnested_or_patterns,
     clippy::option_if_let_else,
@@ -6,42 +11,110 @@
     clippy::cast_sign_loss
 )]
 
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod animation;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod animation;
+#[allow(dead_code)]
 mod app;
+#[allow(dead_code)]
+mod cli;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod config;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod config;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod deletion;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod deletion;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod error;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod error;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod filter;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod filter;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod input;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod input;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod model;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod model;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod native_path;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod native_path;
+#[allow(dead_code)]
 mod os;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod outcome;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod outcome;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod report;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod report;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod runtime;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod runtime;
+#[allow(dead_code)]
 mod state;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod terminal;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod terminal;
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod theme;
+#[cfg(not(any(feature = "fuzzing", feature = "internal")))]
+#[allow(dead_code)]
+mod theme;
+#[allow(dead_code)]
 mod ui;
 #[cfg(windows)]
 mod windows_delete;
 
 pub(crate) use app::{App, UiMode};
-pub use input::TerminalEvents;
-pub use state::FileToDelete;
-pub use terminal::{
-    TerminalSession, TerminalState, TerminalTransition, TerminalTransitionError, validate_terminal,
-};
 
-/// Map layout types.
-///
-/// `Tile` coordinates are terminal columns horizontally and half-rows
-/// vertically, so `HALF_ROWS_PER_CELL` is part of reading a laid-out map.
+/// Internal entry point used by the binary target.
+#[doc(hidden)]
+#[must_use]
+pub fn run_main() -> i32 {
+    cli::run_main()
+}
+
+/// Internal command factory used to regenerate the CLI's distribution artifacts.
+#[doc(hidden)]
+#[must_use]
+pub fn cli_command() -> clap::Command {
+    use clap::CommandFactory as _;
+
+    config::Cli::command()
+}
+
+#[cfg(any(feature = "fuzzing", feature = "internal"))]
 pub mod geometry {
-    /// This summary describes entries that do not fit at the current map size.
-    pub use crate::state::tiles::MapOverflow;
     pub use crate::state::tiles::{FileMetadata, FileType, HALF_ROWS_PER_CELL, Tile, TreeMap};
 }
+#[cfg(feature = "fuzzing")]
+pub use state::FileToDelete;
+#[cfg(feature = "fuzzing")]
+pub use terminal::{TerminalState, TerminalTransition};
 
 #[cfg(test)]
 mod tests;
