+++ xtask/Cargo.toml
@@ -7,7 +7,6 @@ rust-version = "1.88"
 
 
 [dependencies]
-clap = "4.6"
 clap_complete = "4.6"
 clap_mangen = "0.2.31"
 cargo_metadata = "0.23"
